refactor: add hooks
